requirements.txt
# ── Core RAG Stack ────────────────────────────────────────────────────────
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langchain-classic>=0.0.1     # EnsembleRetriever (moved here in LangChain 1.x)

# ── LLM & Embeddings ─────────────────────────────────────────────────────
langchain-google-genai>=1.0.0   # Gemini LLM via Google Generative AI
sentence-transformers>=2.2.0    # Local HuggingFace embeddings (all-MiniLM-L6-v2)
tiktoken>=0.6.0                 # tokenizer used by LangChain text splitter

# ── Vector Database ───────────────────────────────────────────────────────
chromadb>=0.5.0
# pysqlite3-binary>=0.5.0  # Linux/Streamlit Cloud only — not needed on Mac

# ── Advanced Feature: Hybrid Search ──────────────────────────────────────
rank-bm25>=0.2.2         # BM25 index used by BM25Retriever

# ── Document Loaders ─────────────────────────────────────────────────────
pypdf>=4.0.0             # PDF text extraction
beautifulsoup4>=4.12.0   # HTML parsing
lxml>=5.0.0              # HTML parser backend for BeautifulSoup

# ── Environment & Config ──────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Frontend ─────────────────────────────────────────────────────────────
streamlit>=1.32.0

# ── Wikipedia Downloader ──────────────────────────────────────────────────
wikipedia-api>=0.6.0     # used by scripts/download_docs.py

# ── Testing ───────────────────────────────────────────────────────────────
pytest>=7.0.0
